third_party/blink/web_tests/platform/generic/editing/pasteboard/display-block-on-spans-expected.txt

This is a test case for the fix for http://bugzilla.opendarwin.org/show_bug.cgi?id=7904.
ReplaceSelectionCommand used to guess whether or not nodes in the fragment to be pasted were blocks. It now knows for sure.
Before the fix, ReplaceSelectionCommand blew away content of the div below after copy/pasting it over itself.
The following two results must be visually identical except for the caret.

before:
| "
"
| <span>
|   style="display:block"
|   <b>
|     "This is a paragraph."
| "
"
| <span>
|   style="display:block"
|   <b>
|     "This is another paragraph."
| "
"

after:
| "
"
| <span>
|   style="display: block;"
|   <b>
|     "This is a paragraph."
| <span>
|   style="display:block"
|   <b>
|     "This<#selection-caret>"
|   <b>
|     " is another paragraph."
|   <br>
| "
"
